Completed prerequisite
a few functions makes this works
1) In database handler, there is a check if prerequisite exist. It will
search all the past sem to see if the mod exists
2) In the mycalculator there is a function that takes in the
prerequisite string given my nusmods and splits it up to the possible
prerequisite lists. If it is too long it will return nothing because
some mods have junk prerequisite like A-level and some other mods. Most
people fulfil the requirement of the A level and not the other mods

3) This list of random strings will then be send to the database to be
verified if the strings are actually modules or just words

4) the final list would be added into the database

Other changes
1) Timmer for time out of finding the mod is changed to 5 seconds
Bugs:
1) the format of the error message can be changed by changing the
calculator strSeperator
2) When the app is just installed, for the first time, it does not have
the database of the preexisting mods, then it will not function properly
4) Moduleinfo datastructure is also changed to include prereq